% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{as_trackdata}
\alias{as_trackdata}
\title{Format standard Movebank data and add basic movement parameters}
\usage{
as_trackdata(data, min_time_interval = 6)
}
\arguments{
\item{data}{A data.frame of standard Movebank data}
}
\description{
Import standard Movebank data, export data with basic movement parameters
such as distance (between two adjacent points), speed, moving direction, changes of moving direction,
as well as Year, Month, Julian day, Hour of every record.
}
\examples{

attach(movebankdata)
trackdata = as_trackdata(data = movebankdata, min_time_interval = 6)

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
