% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{dist.annual}
\alias{dist.annual}
\title{Calculate annual movement distance (km) for each individual}
\usage{
dist.annual(D = D, min.day = min.day)
}
\arguments{
\item{D}{A data.frame of processed satellite tracking data.}
}
\value{

}
\description{
Sum the annual total movement distance for each individual
}
\examples{

Dist = dist.annual(D=D, min.day=10); Dist

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
