% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{dist.daily}
\alias{dist.daily}
\title{Calculate daily movement distance (km) for each individual}
\usage{
dist.daily(D = D)
}
\arguments{
\item{D}{A data.frame of processed satellite tracking data.}
}
\value{

}
\description{
Sum the daily total movement distance for each individual
}
\examples{

Daily.dist = dist.daily(D); head(Daily.dist)
plot(Daily.dist$Day, Daily.dist$Dist2, col=as.numeric(as.factor(Daily.dist$Individual)), xlab="Julian day", ylab="Flying distance (km)")

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
