% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{dist_annual}
\alias{dist_annual}
\title{Calculate annual movement distance (km) for each individual}
\usage{
dist_annual(trackdata, n = 10)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data.}

\item{n}{The minimum number of days in a year for calculating the total annual movement distance}
}
\description{
Sum the annual total movement distance for each individual
}
\examples{

Dist = dist_annual(trackdata=trackdata, n = 200); Dist

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
