% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{duration}
\alias{duration}
\title{Calculate tracking duration (days) for all individuals}
\usage{
duration(trackdata)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data.}
}
\description{
Calculate tracking duration (from the first day to the last day (or the present day)) for all individuals
}
\examples{

Dur = duration(trackdata); head(Dur)

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
