% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{format.data}
\alias{format.data}
\title{Format standard Movebank data and add basic movement parameters}
\usage{
\method{format}{data}(D = D, Time.s = Time.s)
}
\arguments{
\item{D}{A data.frame of standard Movebank data}
}
\value{

}
\description{
Import standard Movebank data, export data with basic movement parameters such as distance (between two adjacent points), speed, moving direction, changes of moving direction, as well as Year, Month, Julian day, Hour of every record
}
\examples{

 attach(D)
 data = format.data(D = D, Time.s = '2000-1-1 0:0:0')

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
