% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{getEnvData}
\alias{getEnvData}
\title{Provide pseudo-absent points and derive environmental variables}
\usage{
getEnvData(species, buffer, absence, Envlayers)
}
\arguments{
\item{species}{A data.frame of species occurrences with columns "Lon", "Lat", "Time"}

\item{buffer}{A value of distance (unit: degree) defining the width of buffer zone outside the occurrences}

\item{absence}{The number of pseudo-absent points on the longitude side and the latitude side}

\item{Envlayers}{A RasterBrick (a multi-layer raster object)}
}
\value{
Return a data.frame with values of environmental variables at occurrences and pseudo-absent points.
}
\description{
This function provides pseudo-absent points using user defined sample size,
derives values of environmental variables at occurrences and pseudo-absent points.
}
\examples{
 attach(kiang) # load occurrences data
 attach(BioClim)
 Data = getEnvDatakiang, buffer=0.5, absence=30, Envlayers=BioClim); head(Data)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
