% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{HOSDM}
\alias{HOSDM}
\title{Run hetero-occurrence species distribution models (HOSDMs)}
\usage{
HOSDM(
  trackdata = trackdata,
  prediction = F,
  buffer = 0.1,
  absence = 30,
  Envlayer = BioClim
)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data with columns "Lon", "Lat", "Time"#'}

\item{prediction}{A boolean value (T/F) that signifies whether habitat suitability prediction is being performed}

\item{buffer}{define the buffer distance (in degree) for the extent of layers of environmental variables}

\item{absence}{The number of pseudo-absence points are absence*absence}

\item{Envlayer}{A RasterBrick (a multi-layer raster object) for environmental variables}

\item{species}{A subset of trackdata}
}
\description{
HOSDM is a novel species distribution modeling approach that is specifically designed to distinguish
between various types of occurrences. Since White-naped Cranes rely on two separate habitat types, croplands and water areas,
for their daily activities, we categorized the occurrences into two groups and executed HOSDMs accordingly.
}
\examples{
# The number of points during the foraging stage (Points1) and the roosting stage (Points0), respectively.
# The R square values for species distribution models during the two stages (R0 and R1).
# The mean distances between adjacent points during the foraging stage (Dist1) and the roosting stage (Dist0).
# The distance between the centers of the forage area and the roost area (Dist).
results = HOSDM(trackdata = HOSDMdata, prediction=F, buffer=0.1, absence=30, Envlayer=BioClim)
results[[1]] 

# Predicted habitat suitability
plot(results[[2]][[1]], main=paste("Roosting habitat for individual", ind[1], sep=" " ))
plot(results[[3]][[1]], main=paste("Foraging habitat for individual", ind[1], sep=" " ))

}
\author{
Xinhai Li (xinhai_li_edu@126.com)
}
