% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{mig.day}
\alias{mig.day}
\title{Determine the starting day and ending day of migration}
\usage{
mig.day(D = D, min.dist = min.dist)
}
\arguments{
\item{D}{A data.frame of processed satellite tracking data.}
}
\value{

}
\description{
If the daily movement distance of an individual is over a threshold (i.e. 100km), the day is defined as the starting day. The last day of a series of continuous days that movement distance over the threshold is the ending day.
}
\examples{

dates = mig.day(D, min.dist=100) ## minimum distance 100 km

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
