% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{mig.hour}
\alias{mig.hour}
\title{Determine the starting hour and ending hour on the starting and ending day of migration}
\usage{
mig.hour(
  D = D,
  Mig.day = Mig.day,
  outlier.dist = outlier.dist,
  min.move = min.move
)
}
\arguments{
\item{D}{A data.frame of processed satellite tracking data.}

\item{Mig.day}{A data.frame returned from function mig.day().}
}
\value{

}
\description{
If the hourly movement distance of an individual is over a threshold (i.e. 5km), the hour is defined as the starting hour. The ending our is define as the last hour of a series of continuous movement hours in the last day.
}
\examples{

DD = mig.hour(D=D, Mig.day = dates, outlier.dist = 150, min.move = 5)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
