% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{mig_day}
\alias{mig_day}
\title{Determine the starting day and ending day of migration}
\usage{
mig_day(trackdata = trackdata, min.dist = min.dist)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data.}

\item{min.dist}{The threshold movement distance (e.g. 100km / day) for determining the start of a migration.}
}
\value{

}
\description{
If the daily movement distance of an individual is over a threshold (i.e. 100km), the day is defined as the starting day. The last day of a series of continuous days that movement distance over the threshold is the ending day.
}
\examples{

dates = mig_day(trackdata, min.dist=100) ## minimum distance 100 km

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
