% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{mig_hour}
\alias{mig_hour}
\title{Determine the starting hour and ending hour on the starting and ending day of migration}
\usage{
mig_hour(
  trackdata = trackdata,
  Mig.day = Mig.day,
  outlier.dist = outlier.dist,
  min.move = min.move
)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data.}

\item{Mig.day}{A data.frame returned from function mig_day().}

\item{outlier.dist}{The threshold distance (e.g. 150 km) for determining an outlier (caused by missing records). If the distance > 150 km, it is not likely the distance that the bird flied in an hour.}

\item{min.move}{The threshold distance (e.g. 10 km) for determining the start time of migration. Whenever a bird flies over 10 km in an hour, it starts migration.}
}
\value{

}
\description{
If the hourly movement distance of an individual is over a threshold (i.e. 5km), the hour is defined as the starting hour. The ending our is define as the last hour of a series of continuous movement hours in the last day.
}
\examples{

DD = mig_hour(trackdata=trackdata, Mig.day = dates, outlier.dist = 150, min.move = 10)

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
