% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{mig_timing}
\alias{mig_timing}
\title{Determine the starting day and ending day of migration}
\usage{
mig_timing(
  trackdata,
  dist_min_day = 100,
  dist_min_hour = 10,
  dist_outlier = 150
)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data.}

\item{dist_min_day}{The threshold movement distance (e.g. 100km / day) for determining the start date of a migration.}

\item{dist_min_hour}{The threshold movement distance (e.g. 10km / day) for determining the start hour of a migration.}

\item{dist_outlier}{If the movement distance within one hour is over 150km, it is defined as an outlier (caused by missing records).}
}
\description{
If the daily movement distance of an individual is over a threshold (i.e. 100km), the day is defined as the starting day.
The last day of a series of continuous days that movement distance over the threshold is the ending day. Similarly, if the movement distance
in one hour on the migration day (starting or ending day) is over a threshold (i.e. 10km), the "hour" is the starting or ending hour of the movement.
}
\examples{

timing = mig_timing(trackdata=trackdata, dist_min_day = 100, dist_min_hour = 10, dist_outlier = 150)

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
