% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{nest.locating}
\alias{nest.locating}
\title{Identify nest site for each individual}
\usage{
nest.locating(D = D, breed.S = breed.S, breed.E = breed.E, minimum.rec = 100)
}
\arguments{
\item{D}{A data.frame of processed satellite tracking data.}
}
\value{

}
\description{
Determie the nest site for each individual based on the most used site in the breeding season
}
\examples{

nest.locating(D, breed.S = 130, breed.E = 180, minimum.rec = 100)

# Check a single individual
list = names(table(D$ID))
i=1; YEAR=2018
Ind = D[D$ID == list[i] & D$Year==YEAR, ]
breed.S = 140; breed.E = 180
Ind = Ind[Ind$Day > breed.S & Ind$Day < breed.E, ]
plot(Ind$Lon, Ind$Lat)
lines(Ind$Lon, Ind$Lat, col="grey", lwd=.5)
LAT = round(Ind$Lat, 4); LON = round(Ind$Lon, 4)
LATLON = as.character(LAT*LON*10^8) # numeric would cause no-match
frq = sort(table(LATLON), decreasing=T)
LAT = Ind$Lat[LATLON ==  names(frq [frq==max(frq)] )][1]
LON = Ind$Lon[LATLON ==  names(frq [frq==max(frq)] )][1]
points(LON, LAT, col=2, pch=16)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
