% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{nest_locating}
\alias{nest_locating}
\title{Identify nest site for each individual}
\usage{
nest_locating(
  trackdata = trackdata,
  breed.S = breed.S,
  breed.E = breed.E,
  minimum.rec = 100
)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data.}

\item{breed.S}{The starting day of the breeding season}

\item{breed.E}{The ending day of the breeding season}

\item{minimum.rec}{The minimum number needed for estimating nest site.}
}
\value{

}
\description{
Determie the nest site for each individual based on the most used site in the breeding season
}
\examples{

nest_locating(trackdata, breed.S = 130, breed.E = 180, minimum.rec = 50)

# Check a single individual
IDs = names(table(trackdata$ID))
i=1; YEAR=2018
Ind = trackdata[trackdata$ID == IDs[i] & trackdata$Year==YEAR, ]
breed.S = 140; breed.E = 180
Ind = Ind[Ind$Day > breed.S & Ind$Day < breed.E, ]
plot(Ind$Lon, Ind$Lat)
lines(Ind$Lon, Ind$Lat, col="grey", lwd=.5)
LAT = round(Ind$Lat, 4); LON = round(Ind$Lon, 4)
LATLON = as.character(LAT*LON*10^8) # numeric would cause no-match
frq = sort(table(LATLON), decreasing=T)
LAT = Ind$Lat[LATLON ==  names(frq [frq==max(frq)] )][1]
LON = Ind$Lon[LATLON ==  names(frq [frq==max(frq)] )][1]
points(LON, LAT, col=2, pch=16)

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
