% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{plot.breeding.wintering}
\alias{plot.breeding.wintering}
\title{Plot breeding and wintering areas}
\usage{
plot.breeding.wintering(
  D = D,
  ext.par = 4,
  breed.start = breed.start,
  breed.end = breed.end,
  winter.start = winter.start,
  winter.end = winter.end,
  lat.min.b = lat.min.b,
  lat.max.w = lat.max.w,
  breed.percent = c(90, 70, 50),
  winter.percent = c(90, 70, 50)
)
}
\arguments{
\item{D}{A data.frame of processed satellite tracking data.}
}
\value{

}
\description{
Plot breeding and wintering areas based on kernel density maps. The starting and ending dates of the breeding peroid and wintering period are needed. The areas can be adjusted by the percentage of kernel areas.
}
\examples{

 plot.breeding.wintering(D=data, ext.par=4, breed.start=100,
   breed.end=210, winter.start=1, winter.end=60,
   lat.min.b = 45, lat.max.w = 25,
   breed.percent = c(90, 70, 50), winter.percent = c(99.9, 90, 85))

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
