% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{plot.daily.dist}
\alias{plot.daily.dist}
\title{Plot daily movement distance (km) for each individual}
\usage{
plot.daily.dist(D = D)
}
\arguments{
\item{D}{A data.frame returned from function daily.dist().}
}
\value{

}
\description{
Plot the daily total movement distance for each individual in the range of Julian day (1-365)
}
\examples{

Daily.dist = daily.dist(D)
plot.daily.dist(Daily.dist)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
