% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{plot.direction}
\alias{plot.direction}
\title{Plot moving directions of an individual in a year}
\usage{
plot.direction(ind = ind)
}
\arguments{
\item{ind}{Locating records of an individual in a year, from the processed satellite tracking data.}
}
\value{

}
\description{
Use calcualted moving directions by function format.data() to plot moving directions of an individual in a year
}
\examples{

ind = D[D$ID=="钃戠窘楣愼㸴06-683路BFU076" & D$Year == 2018,]
plot.direction(ind)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
