% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{plot.track.duration}
\alias{plot.track.duration}
\title{Plot tracking duration (days) for all individuals}
\usage{
plot.track.duration(D = D)
}
\arguments{
\item{D}{A data.frame of processed satellite tracking data.}
}
\value{

}
\description{
Plot tracking duration (from the first day to the last day (or the present day)) for all individuals
}
\examples{

track.duration.plot(D)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
