% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{plot.traj}
\alias{plot.traj}
\title{Plot trajectories of all individuals in all years}
\usage{
plot.traj(
  D = D,
  ext.par = ext.par,
  transparent = 0.1,
  p.size = 0.2,
  l.size = 1,
  type = "ind"
)
}
\arguments{
\item{D}{A data.frame of processed satellite tracking data.}
}
\value{

}
\description{
Plot two types of trajectories of all individuals in all years. Type "ind" shows different individuals in different color. Type "time" shows locating points in different colors determined by Julian days.
}
\examples{

plot.traj(D=D, ext.par=4, transparent = 0.2, p.size = 0.5, l.size = 1, type="ind")
plot.traj(D=D, ext.par=4, transparent = 0.2, p.size = 0.5, l.size = 1, type="time")

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
