% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{plot_breeding_wintering}
\alias{plot_breeding_wintering}
\title{Plot breeding and wintering areas}
\usage{
plot_breeding_wintering(
  trackdata = trackdata,
  ext.par = 4,
  breed.start = 90,
  breed.end = 210,
  winter.start = 1,
  winter.end = 60,
  lat.min.b = -70,
  lat.max.w = 90,
  breed.percent = c(90, 70, 50),
  winter.percent = c(90, 70, 50)
)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data.}

\item{ext.par}{The buffer distance (degree) outside the species occurrences ranges}

\item{breed.start}{The starting day (Julian day) of breeding migration}

\item{breed.end}{The ending day (Julian day) of breeding migration}

\item{winter.start}{The starting day (Julian day) of wintering migration}

\item{winter.end}{The ending day (Julian day) of wintering migration}

\item{lat.min.b}{The minimum latitude in the breeding season}

\item{lat.max.w}{The maximum latitude in the wintering season}

\item{breed.percent}{A vector of three percentage values defining the kernel density areas in the breeding season}

\item{winter.percent}{A vector of three percentage values defining the kernel density areas in the wintering season}
}
\description{
Plot breeding and wintering areas based on kernel density maps. The starting and ending dates
of the breeding peroid and wintering period are needed. The areas can be adjusted by the percentage of kernel areas.
}
\examples{

 plot_breeding_wintering(trackdata=trackdata, ext.par=4, breed.start=100,
   breed.end=210, winter.start=1, winter.end=60,
   lat.min.b = 45, lat.max.w = 25,
   breed.percent = c(90, 70, 50), winter.percent = c(99.9, 90, 85))

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
