% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{plot_daily_dist}
\alias{plot_daily_dist}
\title{Plot daily movement distance (km) for each individual}
\usage{
plot_daily_dist(data = Daily.dist)
}
\arguments{
\item{trackdata}{A data.frame returned from function daily_dist().}
}
\description{
Plot the daily total movement distance for each individual in the range of Julian day (1-365)
}
\examples{

Daily.dist = dist_daily(trackdata)
plot_daily_dist(Daily.dist)

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
