% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{plot_direction}
\alias{plot_direction}
\title{Plot moving directions of an individual in a year}
\usage{
plot_direction(ind = ind)
}
\arguments{
\item{ind}{Locating records of an individual in a year, from the processed satellite tracking data.}
}
\value{

}
\description{
Use calcualted moving directions by function format.data() to plot moving directions of an individual in a year
}
\examples{

ind = trackdata[trackdata$ID=="钃戠窘楣愼㸴06-683路BFU076" & trackdata$Year == 2018,]
plot_direction(ind)

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
