% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{plot_track_duration}
\alias{plot_track_duration}
\title{Plot tracking duration (days) for all individuals}
\usage{
plot_track_duration(trackdata = trackdata, cex.lab = 1, cex.axis = 1)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data.}
}
\description{
Plot tracking duration (from the first day to the last day (or the present day)) for all individuals
}
\examples{

plot_track_duration(trackdata)

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
