% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{plot_traj}
\alias{plot_traj}
\title{Plot trajectories of all individuals in all years}
\usage{
plot_traj(
  trackdata,
  type = "individual",
  buffer = 4,
  transparent = 0.1,
  p.size = 0.2,
  l.size = 1
)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data.}

\item{type}{There are two options for the type. "ind" shows different individuals in different colors. "time" shows the points from Julian day 1 to 365 by gradient colors red, yellow and green.}

\item{transparent}{The tranparency of colors for points}

\item{p.size}{The size of points}

\item{l.size}{The width of lines}

\item{ext.par}{The buffer distance (degree) outside the species occurrences ranges}
}
\value{

}
\description{
Plot two types of trajectories of all individuals in all years. Type "ind" shows different individuals in different color. Type "time" shows locating points in different colors determined by Julian days.
}
\examples{

plot_traj(trackdata = data, type = "individual")
plot_traj(trackdata = data, type = "chronic")

}
\author{
Huidong Tian (tienhuitung@gmail.com); Xinhai Li (xinhai_li_edu@126.com)
}
