% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{plot_traj_segments}
\alias{plot_traj_segments}
\title{Plot movement trajectories of an individual at any given number of periods}
\usage{
plot_traj_segments(data = ind, seg = 9, label = T)
}
\arguments{
\item{seg}{the number of trajectory segments}

\item{label}{a boolean value determining whether plot the labels of days}

\item{ind}{Locating records of an individual in a year, from the processed satellite tracking data.}
}
\value{

}
\description{
Use calcualted moving directions by function format.data() to plot moving directions of an individual in a year
}
\examples{

ind = trackdata[trackdata$ID==trackdata$ID[1] & trackdata$Year == trackdata$Year[1],]
plot_traj_segments(data=ind, seg=6, label=F)

}
\author{
Xinhai Li (xinhai_li_edu@126.com)
}
