% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationR functions.R
\name{speed_mixedmodel}
\alias{speed_mixedmodel}
\title{Compare movement speed between breeding season and wintering season}
\usage{
speed_mixedmodel(
  trackdata = trackdata,
  breed.start = 150,
  breed.end = 180,
  winter.start = 1,
  winter.end = 31
)
}
\arguments{
\item{trackdata}{A data.frame of processed satellite tracking data.}

\item{breed.start}{The starting day (Julian day) of breeding migration}

\item{breed.end}{The ending day (Julian day) of breeding migration}

\item{winter.start}{The starting day (Julian day) of wintering migration}

\item{winter.end}{The ending day (Julian day) of wintering migration}
}
\description{
Apply a mixed effect model to compare movement speed at breeding season and wintering season, using hour (24h)
as covariate (linear term, quadratic term and interaction term are included). The results show the t test table of fixed variables and terms,
and Marginal R square (based on fixed effect) and Conditional R square (based on both fixed and random effects).
}
\examples{

speed_mixedmodel(trackdata=trackdata, breed.start = 150, breed.end = 180, winter.start=1, winter.end=31)

}
\author{
Xinhai Li (xinhai_li_edu@126.com)
}
