% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane migrationR functions.R
\name{duration}
\alias{duration}
\title{Calculate tracking duration (days) for all individuals}
\usage{
duration(D = D)
}
\arguments{
\item{D}{A data.frame of processed satellite tracking data.}
}
\value{

}
\description{
Calculate tracking duration (from the first day to the last day (or the present day)) for all individuals
}
\examples{

Dur = duration(D); head(Dur)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
